Split REST controller into multiple files
