--- src/ProjectOrigin.WalletSystem.Server/Services/REST/v1/WalletController.cs
@@ -1,253 +1,0 @@
-using System;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using ProjectOrigin.WalletSystem.Server.Database;
-using ProjectOrigin.WalletSystem.Server.Extensions;
-using ProjectOrigin.WalletSystem.Server.Models;
-
-namespace ProjectOrigin.WalletSystem.Server.Services.REST.v1;
-
-[Authorize]
-[ApiController]
-public class WalletController : ControllerBase
-{
-    /// <summary>
-    /// Gets all certificates in the wallet that are <b>available</b> for use.
-    /// </summary>
-    /// <param name="unitOfWork"></param>
-    /// <param name="start">The start of the time range in Unix time in seconds.</param>
-    /// <param name="end">The end of the time range in Unix time in seconds.</param>
-    /// <response code="200">Returns the aggregated claims.</response>
-    /// <response code="401">If the user is not authenticated.</response>
-    [HttpGet]
-    [Route("v1/certificates")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<ResultList<GranularCertificate>>> GetCertificates([FromServices] IUnitOfWork unitOfWork, [FromQuery] long? start, [FromQuery] long? end)
-    {
-        if (!User.TryGetSubject(out var subject)) return Unauthorized();
-
-        var certificates = await unitOfWork.CertificateRepository.GetAllOwnedCertificates(subject, new CertificatesFilter
-        {
-            Start = start != null ? DateTimeOffset.FromUnixTimeSeconds(start.Value) : null,
-            End = end != null ? DateTimeOffset.FromUnixTimeSeconds(end.Value) : null
-        });
-
-        return new ResultList<GranularCertificate> { Result = certificates.Select(c => c.MapToV1()) };
-    }
-
-    /// <summary>
-    /// Gets all claims in the wallet
-    /// </summary>
-    /// <param name="unitOfWork"></param>
-    /// <param name="start">The start of the time range in Unix time in seconds.</param>
-    /// <param name="end">The end of the time range in Unix time in seconds.</param>
-    /// <response code="200">Returns all the indiviual claims.</response>
-    /// <response code="401">If the user is not authenticated.</response>
-    [HttpGet]
-    [Route("v1/claims")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<ResultList<Claim>>> GetClaims([FromServices] IUnitOfWork unitOfWork, [FromQuery] long? start, [FromQuery] long? end)
-    {
-        if (!User.TryGetSubject(out var subject)) return Unauthorized();
-
-        var claims = await unitOfWork.CertificateRepository.GetClaims(subject, new ClaimFilter
-        {
-            Start = start != null ? DateTimeOffset.FromUnixTimeSeconds(start.Value) : null,
-            End = end != null ? DateTimeOffset.FromUnixTimeSeconds(end.Value) : null,
-        });
-
-        return new ResultList<Claim> { Result = claims.Select(c => c.MapToV1()) };
-    }
-
-    /// <summary>
-    /// Gets detailed list of all of the transfers that have been made to other wallets.
-    /// </summary>
-    /// <param name="unitOfWork"></param>
-    /// <param name="start">The start of the time range in Unix time in seconds.</param>
-    /// <param name="end">The end of the time range in Unix time in seconds.</param>
-    /// <response code="200">Returns the individual transferes within the filter.</response>
-    /// <response code="401">If the user is not authenticated.</response>
-    [HttpGet]
-    [Route("v1/transfers")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<ResultList<Transfer>>> GetTransfers([FromServices] IUnitOfWork unitOfWork, [FromQuery] long? start, [FromQuery] long? end)
-    {
-        if (!User.TryGetSubject(out var subject)) return Unauthorized();
-
-        var transfers = await unitOfWork.CertificateRepository.GetTransfers(subject, new TransferFilter
-        {
-            Start = start != null ? DateTimeOffset.FromUnixTimeSeconds(start.Value) : null,
-            End = end != null ? DateTimeOffset.FromUnixTimeSeconds(end.Value) : null,
-        });
-
-        return new ResultList<Transfer>
-        {
-            Result = transfers.Select(t => new Transfer()
-            {
-                FederatedStreamId = new FederatedStreamId
-                {
-                    Registry = t.RegistryName,
-                    StreamId = t.CertificateId
-                },
-                ReceiverId = t.ReceiverId.ToString(),
-                Start = t.StartDate.ToUnixTimeSeconds(),
-                End = t.EndDate.ToUnixTimeSeconds(),
-                Quantity = t.Quantity,
-                GridArea = t.GridArea,
-            })
-        };
-    }
-
-    /// <summary>
-    /// Returns aggregates certificates that are <b>available</b> to use, based on the specified time zone and time range.
-    /// </summary>
-    /// <param name="unitOfWork"></param>
-    /// <param name="timeAggregate">The size of each bucket in the aggregation</param>
-    /// <param name="timeZone">The time zone. See https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for a list of valid time zones.</param>
-    /// <param name="start">The start of the time range in Unix time in seconds.</param>
-    /// <param name="end">The end of the time range in Unix time in seconds.</param>
-    /// <param name="type">Filter the type of certificates to return.</param>
-    /// <response code="200">Returns the aggregated claims.</response>
-    /// <response code="400">If the time zone is invalid.</response>
-    /// <response code="401">If the user is not authenticated.</response>
-    [HttpGet]
-    [Route("v1/aggregate-certificates")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<ResultList<AggregatedCertificates>>> AggregateCertificates(
-        [FromServices] IUnitOfWork unitOfWork,
-        [FromQuery] TimeAggregate timeAggregate,
-        [FromQuery] string timeZone,
-        [FromQuery] long? start,
-        [FromQuery] long? end,
-        [FromQuery] CertificateType? type)
-    {
-        if (!User.TryGetSubject(out var subject)) return Unauthorized();
-        if (!timeZone.TryParseTimeZone(out var timeZoneInfo)) return BadRequest("Invalid time zone");
-
-        var certificates = await unitOfWork.CertificateRepository.GetAllOwnedCertificates(subject, new CertificatesFilter
-        {
-            Start = start != null ? DateTimeOffset.FromUnixTimeSeconds(start.Value) : null,
-            End = end != null ? DateTimeOffset.FromUnixTimeSeconds(end.Value) : null,
-            Type = type != null ? (GranularCertificateType)type.Value : null
-        });
-
-        return new ResultList<AggregatedCertificates>
-        {
-            Result = certificates
-                .GroupBy(cert => cert.CertificateType)
-                .SelectMany(typeGroup => typeGroup
-                    .GroupByTime(x => x.StartDate, (Models.TimeAggregate)timeAggregate, timeZoneInfo)
-                    .Select(timeGroup => new AggregatedCertificates
-                    {
-                        Type = (CertificateType)typeGroup.Key,
-                        Quantity = timeGroup.Sum(certificate => certificate.Slices.Sum(slice => slice.Quantity)),
-                        Start = timeGroup.Min(certificate => certificate.StartDate).ToUnixTimeSeconds(),
-                        End = timeGroup.Max(certificate => certificate.EndDate).ToUnixTimeSeconds(),
-                    }))
-        };
-    }
-
-    /// <summary>
-    /// Returns a list of aggregates claims for the authenticated user based on the specified time zone and time range.
-    /// </summary>
-    /// <param name="unitOfWork"></param>
-    /// <param name="timeAggregate">The size of each bucket in the aggregation</param>
-    /// <param name="timeZone">The time zone. See https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for a list of valid time zones.</param>
-    /// <param name="start">The start of the time range in Unix time in seconds.</param>
-    /// <param name="end">The end of the time range in Unix time in seconds.</param>
-    /// <response code="200">Returns the aggregated claims.</response>
-    /// <response code="400">If the time zone is invalid.</response>
-    /// <response code="401">If the user is not authenticated.</response>
-    [HttpGet]
-    [Route("v1/aggregate-claims")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<ResultList<AggregatedClaims>>> AggregateClaims(
-        [FromServices] IUnitOfWork unitOfWork,
-        [FromQuery] TimeAggregate timeAggregate,
-        [FromQuery] string timeZone,
-        [FromQuery] long? start,
-        [FromQuery] long? end)
-    {
-        if (!User.TryGetSubject(out var subject)) return Unauthorized();
-        if (!timeZone.TryParseTimeZone(out var timeZoneInfo)) return BadRequest("Invalid time zone");
-
-        var claims = await unitOfWork.CertificateRepository.GetClaims(subject, new ClaimFilter
-        {
-            Start = start != null ? DateTimeOffset.FromUnixTimeSeconds(start.Value) : null,
-            End = end != null ? DateTimeOffset.FromUnixTimeSeconds(end.Value) : null,
-        });
-
-        return new ResultList<AggregatedClaims>
-        {
-            Result = claims
-                .GroupByTime(x => x.ProductionStart, (Models.TimeAggregate)timeAggregate, timeZoneInfo)
-                .Select(group => new AggregatedClaims
-                {
-                    Quantity = group.Sum(claim => claim.Quantity),
-                    Start = group.Min(claim => claim.ProductionStart).ToUnixTimeSeconds(),
-                    End = group.Max(claim => claim.ProductionEnd).ToUnixTimeSeconds(),
-                })
-        };
-    }
-
-    /// <summary>
-    /// Returns a list of aggregates transfers, for all certificates transferred to another wallet for the authenticated user based.
-    /// </summary>
-    /// <param name="unitOfWork"></param>
-    /// <param name="timeAggregate">The size of each bucket in the aggregation</param>
-    /// <param name="timeZone">The time zone. See https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for a list of valid time zones.</param>
-    /// <param name="start">The start of the time range in Unix time in seconds.</param>
-    /// <param name="end">The end of the time range in Unix time in seconds.</param>
-    /// <response code="200">Returns the aggregated claims.</response>
-    /// <response code="400">If the time zone is invalid.</response>
-    /// <response code="401">If the user is not authenticated.</response>
-    [HttpGet]
-    [Route("v1/aggregate-transfers")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(void), StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<ResultList<AggregatedTransfers>>> AggregateTransfers(
-        [FromServices] IUnitOfWork unitOfWork,
-        [FromQuery] TimeAggregate timeAggregate,
-        [FromQuery] string timeZone,
-        [FromQuery] long? start,
-        [FromQuery] long? end)
-    {
-        if (!User.TryGetSubject(out var subject)) return Unauthorized();
-        if (!timeZone.TryParseTimeZone(out var timeZoneInfo)) return BadRequest("Invalid time zone");
-
-        var transfers = await unitOfWork.CertificateRepository.GetTransfers(subject, new TransferFilter
-        {
-            Start = start != null ? DateTimeOffset.FromUnixTimeSeconds(start.Value) : null,
-            End = end != null ? DateTimeOffset.FromUnixTimeSeconds(end.Value) : null,
-        });
-
-        return new ResultList<AggregatedTransfers>
-        {
-            Result = transfers
-                .GroupByTime(record => record.StartDate, (Models.TimeAggregate)timeAggregate, timeZoneInfo)
-                .Select(group => new AggregatedTransfers()
-                {
-                    Quantity = group.Sum(transfer => transfer.Quantity),
-                    Start = group.Min(transfer => transfer.StartDate).ToUnixTimeSeconds(),
-                    End = group.Max(transfer => transfer.EndDate).ToUnixTimeSeconds(),
-                })
-        };
-    }
-}
